moved the tests
